turning weights into floats and adding a test with real data

+++ httomolibgpu/misc/morph.py
@@ -84,7 +84,7 @@ def __sino_360_to_180(
     out = cp.empty((n, dy, 2 * dz - overlap), dtype=data.dtype)
 
     if rotation == "left":
-        weights = cp.linspace(0, 1.0, overlap)
+        weights = cp.linspace(0, 1.0, overlap, dtype=cp.float32)
         out[:, :, -dz + overlap :] = data[:n, :, overlap:]
         out[:, :, : dz - overlap] = data[n : 2 * n, :, overlap:][:, :, ::-1]
         out[:, :, dz - overlap : dz] = (
